zdsp1d: Start experiments on running full server under test
Problem on zdsp1d: We have to change the whole SCPI implementation. To get a
regression test it would be helpful to fully mock zdsp1d so we can step by step
move to new implementation.
It might turn into a mission impossible. If so then we have to do error prone
smoke tests on device :(

So let's see how far we get and maybe we learn to mock other services more.
Remember: The mock services we have so far 'just' support announcing resources
and make them questionable by SCPI.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>