com5003d/mt310s2d: Stop naming variables for different settings same
It took us hours to fix/understand a regression we accidentally introduced on
mocking atmel watcher.
It was caused by using variable name 'm_pFPGASettings' on com5003 for
fpga-ctrl and on mt310s2 for (not yet used) message device.

To get rid of this pitfall, rename variables (classes used are still misleading)

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>