Com5003SenseInterface: Reduce dependencies by injecting only what's necessary
We had to learn the hard way that modulemanager needs to be created before
passing its pointer along.

This is another painful compromise: On the one hand we pass yet another bare
pointer and assume 'we are a server and live forever'. On the other hand this
is currently the only way to get Com5003SenseInterface under test. Trying this
with full server in the bag is a mission impossible.
We have to take smaller steps first: Mocking Atmel step by step is easier to
control than mocking all in one step.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>