com5003: Add tests for range constant values
Let's remember where this mission started:
Our target was to find out differences between implementation of SCPI interface
in COM5003 / MT310s2. Although we almost forgot during the jorney, we reached
that target.

Now that we have informational data, a reliable test harness and learned a lot,
we are better off merging the two implementations to one.

Next step should be:
Check if our assumption that rejection values are used only on range module to
decide on overload conditions and no measurement accuracy is affected.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>