com5003justdata: Carefully sync to mt310s2justdata
Called functions do same as code replaced. For full sync we need to add tests
to see if our assumption is correct that we loose!? accuracy by limiting
digits. Possible explanation: mt310s2 was changed to not limiting and com was
forgotten.

In case assumption is tested valid last steps on sync are:

* sync Com5003JustRangeTripletOffsetGainPhase::executeProtoScpi
* sync Com5003JustRangeTripletOffsetGainPhase::scpiQuery*CorrectionTotal to
  those in Com5003JustRangeTripletOffsetGainPhase

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>