Mt310s2AdjXML: Cleanup
Our current plan is to use Mt310s2AdjXML as a blueprint for Com5003 to get
functionality moved from Com5003Adjustment to Com5003SenseInterface.
It is not that we would call the code around Mt310s2AdjXML 'design') but we
need small steps and try to follow the path MT310s2 went through diverging
from COM5003.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>