I2c controllers common info: Start on Demo/Test prefixes
We want to enhance test common info. Currently all members return 'Unknown'
which might lead to wrong links still passing tests.
TestI2cCtrlCommonInfo still reports 'Unknown' because we expect many tests
fail: We recorded adjustment regression data with 'Unknown' because that is
default on SystemInfo when I2c is not working which it was when we started the
regression tests.
Ahh and zera-classes' test_scpi_cmds_in_session::initialScpiCommandsOnOtherModules
is already broken so do some consolidation here before fixing it there for the
last time hopefully.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>