zdsp1d: Add command to remove command lists for all clients in one transaction
We did some profiling on zera-classes / test_modman_regression_all_sessions to
analyze why session change is so terribly slow even on tests.

That was kind of enlightenment: When shutting down modules, each module removes
itself from DSP command list which causes a full command list compile/upload
for the remaining modules which is a terrible idea when it comes to
performance.

So introduce a command to remove the whole DSP command list in one transaction.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>