zdspserver: Add query how many clients are kept in warehouses
There are three containers (one list / two hashes) and it is almost
impossible to find out which is responsible for what and if they are kept
in sync (we guess they should be synced).

So get counts of all and write tests to learn how they work and how we can
replace this obscurity by something readable.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>