adjustment: Prepare compute for ranges that are worth computing
In zera-classes/adjustment-module we recently made compute SCPI deferred to not
miss error cases.
Tests on device showed that compute now is taking several seconds to finish.
This is caused by computing all ranges and - before doing so - asking
microcontroller for permission. Try to avoid that by just computing ranges that
were changed.
For now AdjDataItem::isComputeRequired() returns true so now change of behavior is
expected.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>